Fully functional version for a single iteration. Multiple iterations tested manually.

+++ mg_blast_parser_v1.5.py
@@ -60,6 +60,13 @@ def unique_scaffold_topEval(dataframe):
     #variables=['quid', 'suid', 'iden', 'alen', 'mism', 'gapo', 'qsta', 'qend', 'ssta', 'send', 'eval', 'bits']
     df = pandas.DataFrame([[getattr(i,j) for j in variables] for i in rows], columns = variables)
     return df
+    
+def close_ind_lst(ind_lst):
+    """
+    Closes index objects supplied in input parameter list
+    """
+    for index in ind_lst:
+        index.close()
 
 def usage():
     print "\nThis is the usage function\n"
@@ -84,17 +91,29 @@ def main(argv):
     results_Dir = ""
     input_files_Dir = ""
     ref_out_0 = ""
+    blasted_lst = []
+    continue_from_previous = False #poorly supported, just keeping the directories
+    skip_blasting = False
     
              
     try:                                
-        opts, args = getopt.getopt(argv, "r:m:n:e:a:i:f:h", ["reference=", "metagenome=", "name=", "e_value=", "alignment_length=", "identity=","format=", "iterations=", "alen_increment=", "iden_increment=", "help"])
+        opts, args = getopt.getopt(argv, "r:m:n:e:a:i:f:h", ["reference=", "metagenome=", "name=", "e_value=", "alignment_length=", "identity=","format=", "iterations=", "alen_increment=", "iden_increment=","continue_from_previous","skip_blasting", "help"])
     except getopt.GetoptError:          
         usage()                         
         sys.exit(2)                     
     for opt, arg in opts:                
         if opt in ("-h", "--help"):      
             usage()              
-            sys.exit()                  
+            sys.exit() 
+#        elif opt in ("--recover_after_failure"):
+#            recover_after_failure = True
+#            print "Recover after failure:", recover_after_failure  
+        elif opt in ("--skip_blasting"):
+            skip_blasting = True
+            print "Blasting step omitted; Using previous blast output."
+        elif opt in ("--continue_from_previous"):
+            continue_from_previous = True
+            print "Continue after failure:", continue_from_previous
         elif opt in ("-r", "--reference"):
             if arg:
                 ref_lst=arg.split(',')
@@ -200,13 +219,14 @@ def main(argv):
  
     
     project_dir = name
-    if os.path.exists(project_dir):
-        shutil.rmtree(project_dir)
-    try:
-        os.mkdir(project_dir)
-    except OSError:
-        print "ERROR: Cannot create project directory: " + name
-        raise
+    if not continue_from_previous:
+        if os.path.exists(project_dir):
+            shutil.rmtree(project_dir)
+        try:
+            os.mkdir(project_dir)
+        except OSError:
+            print "ERROR: Cannot create project directory: " + name
+            raise
     
     print "\n\t Initial Parameters:"
     print "\nProject Name: ", name,'\n'
@@ -214,8 +234,8 @@ def main(argv):
     print "Reference File(s): ", ref_lst,'\n'
     print "Metagenome File(s): ", mg_lst,'\n'
     print "E Value: ", e_val, "\n"
-    print "Alignment Length: ", alen,'\n'
-    print "Sequence Identity: ", iden,'\n'
+    print "Alignment Length: "+str(alen)+'%\n'
+    print "Sequence Identity: "+str(iden)+'%\n'
     print "Output Format(s):", fmt_lst,'\n'
     if iterations > 1:
         print "Iterations: ", iterations, '\n'
@@ -224,31 +244,36 @@ def main(argv):
 
     #Initializing directories
     blast_db_Dir = name+"/blast_db"
-    if os.path.exists(blast_db_Dir):
-        shutil.rmtree(blast_db_Dir)
-    try:
-        os.mkdir(blast_db_Dir)
-    except OSError:
-        print "ERROR: Cannot create project directory: " + blast_db_Dir
-        raise
+    if not continue_from_previous:
+        if os.path.exists(blast_db_Dir):
+            shutil.rmtree(blast_db_Dir)
+        try:
+            os.mkdir(blast_db_Dir)
+        except OSError:
+            print "ERROR: Cannot create project directory: " + blast_db_Dir
+            raise
 
     results_Dir = name+"/results"
-    if os.path.exists(results_Dir):
-        shutil.rmtree(results_Dir)
-    try:
-        os.mkdir(results_Dir)
-    except OSError:
-        print "ERROR: Cannot create project directory: " + results_Dir
-        raise
+    if not continue_from_previous:
+    
+        if os.path.exists(results_Dir):
+            shutil.rmtree(results_Dir)
+        try:
+            os.mkdir(results_Dir)
+        except OSError:
+            print "ERROR: Cannot create project directory: " + results_Dir
+            raise
 
     input_files_Dir = name+"/input_files"
-    if os.path.exists(input_files_Dir):
-        shutil.rmtree(input_files_Dir)
-    try:
-        os.mkdir(input_files_Dir)
-    except OSError:
-        print "ERROR: Cannot create project directory: " + input_files_Dir
-        raise
+    if not continue_from_previous:
+    
+        if os.path.exists(input_files_Dir):
+            shutil.rmtree(input_files_Dir)
+        try:
+            os.mkdir(input_files_Dir)
+        except OSError:
+            print "ERROR: Cannot create project directory: " + input_files_Dir
+            raise
 
 # Writing raw reference files into a specific input filename
     input_ref_records = {}
@@ -263,7 +288,117 @@ def main(argv):
     with open(ref_out_0, "w") as handle:
         SeqIO.write(input_ref_records.values(), handle, "fasta")
             #NO NEED TO CLOSE with statement will automatically close the file
+
+# Making BLAST databases
+    #output fname from before used as input for blast database creation
+    input_ref_0 = ref_out_0
+    title_db = name+"_db"#add iteration functionality
+    outfile_db = blast_db_Dir+"/iteration"+str(iterations)+"/"+name+"_db"#change into for loop
+    os.system("makeblastdb -in "+input_ref_0+" -dbtype nucl -title "+title_db+" -out "+outfile_db+" -parse_seqids")
+    
+# BLASTing query contigs
+    if not skip_blasting:
+        print "\nBLASTing query file(s):"
+        for i in range(len(mg_lst)):
+            
+            database = outfile_db # adjust for iterations
+            blasted_lst.append(results_Dir+"/recruited_mg_"+str(i)+".tab")
+            start = time.time()
+            os_string = 'blastn -db '+database+' -query \"'+mg_lst[i]+'\" -out '+blasted_lst[i]+" -evalue "+str(e_val)+"  -outfmt 6 -num_threads 8"
+            #print os_string
+            os.system(os_string)
+            print "\t"+mg_lst[i]+"; Time elapsed: "+str(time.time()-start)+" seconds."
+    else:
+        for i in range(len(mg_lst)):
+            blasted_lst.append(results_Dir+"/recruited_mg_"+str(i)+".tab")
         
+        
+# Parsing BLAST outputs
+    blast_cols = ['quid', 'suid', 'iden', 'alen', 'mism', 'gapo', 'qsta', 'qend', 'ssta', 'send', 'eval', 'bits']
+    recruited_mg=[]
+    for i in range(len(mg_lst)):
+        df = pandas.read_csv(blasted_lst[i] ,sep="\t", header=None)
+        df.columns=blast_cols
+        recruited_mg.append(df)
+        
+#    print len(recruited_mg[0])
+#    print len(recruited_mg[1])
+
+    #creating all_records entry
+#! Remember to close index objects after they are no longer needed
+#! Use helper function close_ind_lst()
+    all_records = []
+    print "\nIndexing metagenome file(s):"
+    for i in range(len(mg_lst)):
+        start = time.time()
+        all_records.append(parse_contigs_ind(mg_lst[i]))
+        print "\t"+mg_lst[i]+" Indexed in : "+str(time.time()-start)+" seconds."
+
+# Transforming data
+    for i in range(len(mg_lst)):
+    #cutoff_contigs[dataframe]=evalue_filter(cutoff_contigs[dataframe])
+        recruited_mg[i]=unique_scaffold_topEval(recruited_mg[i])
+        contig_list = recruited_mg[i]['quid'].tolist()
+        recruited_mg[i]['Seq_nt']=retrive_sequence(contig_list, all_records[i])
+        recruited_mg[i]['Seq_size']=recruited_mg[i]['Seq_nt'].apply(lambda x: len(x))
+        recruited_mg[i]['Coverage']=recruited_mg[i]['alen'].apply(lambda x: float(x))/recruited_mg[i]['Seq_size']
+        recruited_mg[i]['Metric']=recruited_mg[i]['Coverage']*recruited_mg[i]['iden']
+        recruited_mg[i] = recruited_mg[i][['quid', 'suid', 'iden', 'alen','Coverage','Metric', 'mism', 'gapo', 'qsta', 'qend', 'ssta', 'send', 'eval', 'bits','Seq_size', 'Seq_nt']]
+   
+# Here would go statistics functions and producing plots
+#
+#
+#
+#
+#
+
+# Quality filtering before outputting
+    for i in range(len(recruited_mg)):
+        recruited_mg[i]=recruited_mg[i][(recruited_mg[i]['iden']>=iden)&(recruited_mg[i]['Metric']>=alen)&(recruited_mg[i]['eval']<=e_val)]
+
+#    print  len(recruited_mg[0])
+#    print len(recruited_mg[1])
+
+# Batch export to outfmt (csv and/or multiple FASTA)
+    if iterations > 1:
+        prefix=results_Dir+"/"+name+"_iter_e_"+str(e_val)+"_iden_"+str(iden)+"_alen_"+str(alen)+"_recruited_mg_"
+    else:
+        prefix=results_Dir+"/"+name+"_e_"+str(e_val)+"_iden_"+str(iden)+"_alen_"+str(alen)+"_recruited_mg_"
+    
+    print "\nWriting files:"
+    for i in range(len(mg_lst)):
+        records= []
+        
+    #     try:
+    #         os.remove(outfile1)     
+    #     except OSError:
+    #         pass
+        
+        if "csv" in fmt_lst:
+            outfile1 = prefix+str(i)+".csv"
+            recruited_mg[i].to_csv(outfile1, sep='\t')
+            print str(len(recruited_mg[i]))+" sequences written to "+outfile1
+        if "fasta" in fmt_lst:
+            ids = recruited_mg[i]['quid'].tolist()
+            #if len(ids)==len(sequences):
+            for j in range(len(ids)):
+                records.append(all_records[i][ids[j]])
+            outfile2 = prefix+str(i)+".fasta" 
+            with open(outfile2, "w") as output_handle:
+                SeqIO.write(records, output_handle, "fasta")
+            print str(len(ids))+" sequences written to "+outfile2
+    close_ind_lst(all_records)
+    #all_records[i].close()# keep open if multiple iterations
+
+#recruited_mg_1 = pandas.read_csv(out_name1 ,sep="\t", header=None)
+#recruited_mg_1.columns=['quid', 'suid', 'iden', 'alen', 'mism', 'gapo', 'qsta', 'qend', 'ssta', 'send', 'eval', 'bits']
+
+#recruited_mg_2 = pandas.read_csv(out_name2 ,sep="\t", header=None)
+#recruited_mg_2.columns=['quid', 'suid', 'iden', 'alen', 'mism', 'gapo', 'qsta', 'qend', 'ssta', 'send', 'eval', 'bits']
+
+#recruited_mg = [recruited_mg_1, recruited_mg_2]
+
+
 
 
 #    blast_db_Dir = ""
